pyproject.toml
#
# Python build system configuration
# https://peps.python.org/pep-0518/
#

[build-system]
requires      = ["pdm-backend", "versioningit"]
build-backend = "pdm.backend"

[project]
name = "eyrie"
version = "0.0.5"
description = "Your SDLC in a bin."
readme = "README.md"
requires-python = "<4.0,>=3.12"
license = { text = "MIT" }
keywords = []
authors = [{ name = "Ada Wright", email = "ada@hpk.io" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "pyyaml>=6.0.1",
  "click>=8.1.7",
  "versioningit>=3.1.1",
  "pexpect>=4.9.0",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.2.1",
]

[project.scripts]
eyrie-cli = "eyrie.cli:cli"

[project.urls]
Documentation = "https://github.com/hipikat/eyrie#readme"
Issues        = "https://github.com/hipikat/eyrie/issues"
Source        = "https://github.com/hipikat/eyrie"

#
# Coverage - https://coverage.readthedocs.io/
#
[tool.coverage.run]
source_pkgs = ["eyrie", "tests"]
branch      = true
parallel    = true
omit        = ["src/eyrie/__about__.py"]

[tool.coverage.paths]
eyrie = ["src/eyrie", "*/eyrie/src/eyrie"]
tests = ["tests", "*/eyrie/tests"]

[tool.coverage.report]
exclude_lines = ["no cov", "if __name__ == .__main__.:", "if TYPE_CHECKING:"]

#
# PDM (Python Package Management) - https://pdm-project.org/
#
[tool.pdm.build]
includes = ["config/*.yaml"]
# package-dir = "src"

[tool.pdm.dev-dependencies]
dev = ["ipython<9.0.0,>=8.23.0", "pip>=24.0"]

#
# versioningit - https://versioningit.readthedocs.io/
#
[tool.versioningit.format]
# When there have been commits since the most recent tag:
distance = "{version}.dev{distance}+{vcs}{rev}"
# When there are uncommitted changes:
dirty = "{version}+dirty"
# When there are both commits and uncommitted changes:
distance-dirty = "{version}.dev{distance}+{vcs}{rev}.dirty"

[tool.versioningit.vcs]
method      = "git"
match       = ["v*"]
default-tag = "v0.0.0"

[tool.versioningit.tag2version]
method        = "basic"
regex         = "^v(\\d+\\.\\d+\\.\\d+)(?:-([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?(?:\\+([0-9A-Za-z-]+(?:\\.[0-9A-Za-z-]+)*))?$"
require-match = true
